requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
django-cors-headers>=4.0,<5.0
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
celery>=5.3,<6.0
redis>=4.4,<5.0
django-environ>=0.10,<0.11
mysqlclient>=2.2,<3.0
daphne>=4.0,<5.0 
drf-spectacular>=0.27,<0.28 